+++ packages/article-main-comment/src/article-header/article-header.js
@@ -3,6 +3,7 @@ import {
   ArticleFlags,
   UpdatedTimeProvider
 } from "@times-components/ts-components";
+import { InArticleAudio, TCThemeProvider } from "@times-components/ts-newskit";
 import Image from "@times-components/image";
 import { checkStylesForUnits } from "@times-components/utils";
 
@@ -30,6 +31,7 @@ const ArticleHeader = ({
   headline,
   label,
   publicationName,
+  showAudioPlayer,
   publishedTime,
   standfirst,
   updatedTime
@@ -61,6 +63,24 @@ const ArticleHeader = ({
       publicationName={publicationName}
       publishedTime={publishedTime}
     />
+    {showAudioPlayer && (
+      <TCThemeProvider>
+        <InArticleAudio
+          src="https://ncu-newskit-docs.s3.eu-west-1.amazonaws.com/storybook-assets/audio_file_1.mp3"
+          readyToPlayText="Listen to article"
+          playingText="Playing"
+          narrator="James Marriott"
+          headline="Sorry, we can’t all be destined for greatness"
+          feedback={{
+            requestFeedback: true,
+            feedbackMessage:
+              "Want to listen to more articles? Give your feedback below or email",
+            thankyouMessage:
+              "Thank you for your feedback. We're always trying to give you the best possible experience – your feedback helps us do this."
+          }}
+        />
+      </TCThemeProvider>
+    )}
   </HeaderContainer>
 );