+++ packages/ts-newskit/src/components/article-audio/__tests__/stickyplayer.test.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import '@testing-library/jest-dom';
 import { useBreakpointKey } from 'newskit';
-import { render, screen, fireEvent } from '../../../../utils/test-utils';
-import { InArticleAudio, InArticleAudioProps } from '..';
+import { render, screen, fireEvent } from '../../../utils/test-utils';
+import { InArticleAudio, InArticleAudioProps } from '../';
 
 jest.mock('newskit', () => ({
   ...jest.requireActual('newskit'),