+++ packages/ts-newskit/src/components/slices/shared-styles/index.ts
@@ -4,7 +4,11 @@ import {
   Block,
   getSizingCssFromTheme,
   getColorCssFromTheme,
-  CardLink
+  CardLink,
+  CardMedia,
+  getSpacingCssFromTheme,
+  getMediaQueryFromTheme,
+  IconButton
 } from 'newskit';
 
 export const CardHeadlineLink = styled(CardLink)<{ $color?: string }>`
@@ -30,4 +34,18 @@ export const ColouredText = styled(TextBlock)<{ $color?: string }>`
 export const ContainerInline = styled(Block)`
   display: inline-block;
   ${getSizingCssFromTheme('height', 'sizing020')};
+`;
+
+export const TransparentButton = styled(IconButton)`
+  &,
+  &:hover:not(:disabled) {
+    background-color: transparent;
+  }
+`;
+
+const setImageMargin = (space: string) => ({ marginInline: `-${space}` });
+export const FullWidthCardMediaMob = styled(CardMedia)`
+  ${getMediaQueryFromTheme('xs', 'md')} {
+    ${getSpacingCssFromTheme(setImageMargin, 'space045')};
+  }
 `;